services/api-gateway/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pika==1.3.2
psycopg2-binary==2.9.9
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
email-validator
passlib==1.7.4
bcrypt==4.0.1